'payment-by-paypal'

+++ app/src/main/java/com/example/sg772/foodorder/placeOrderActivity.kt
@@ -23,10 +23,12 @@ import android.widget.TextView
 
 import java.util.ArrayList
 import android.Manifest.permission.READ_CONTACTS
+import android.app.Activity
 import android.content.Intent
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
 import android.widget.Button
+import android.widget.Toast
 import com.example.sg772.foodorder.Model.Order
 import com.example.sg772.foodorder.Model.Request
 import com.example.sg772.foodorder.utils.DBHelper
@@ -35,9 +37,12 @@ import com.example.sg772.foodorder.viewHolder.placeOrderList
 import com.google.android.gms.common.internal.service.Common
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.FirebaseDatabase
+import com.paypal.android.sdk.payments.*
 
 import kotlinx.android.synthetic.main.activity_place_order.*
+import org.json.JSONException
 import java.io.Serializable
+import java.math.BigDecimal
 
 /**
  * A login screen that offers login via email/password.
@@ -55,10 +60,23 @@ class placeOrderActivity : AppCompatActivity() {
     lateinit var layoutManager: LinearLayoutManager
     lateinit var orderedFoodAdapter: RequestAdapter
     lateinit var makeOrder: Button
+    companion object {
+        val PAYPAL_REQUEST_CODE: Int=9999
+        var paypalConfig=PayPalConfiguration().environment(PayPalConfiguration.ENVIRONMENT_SANDBOX).clientId(com.example.sg772.foodorder.utils.paypalConfig.clientId)
+    }
+    lateinit var address: String
+    lateinit var comment: String
+
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_place_order)
         // init
+        //paypal
+        var intent = Intent(this, PayPalService::class.java)
+        intent.putExtra(PayPalService.EXTRA_PAYPAL_CONFIGURATION, paypalConfig)
+        startService(intent)
+        //
         Name = findViewById(R.id.place_order_name)
         Phone = findViewById(R.id.place_order_phone)
         Address = findViewById(R.id.place_order_address)
@@ -66,6 +84,11 @@ class placeOrderActivity : AppCompatActivity() {
         //recycler view
         var db = DBHelper(this)
         listOrdered = db.readData()
+        var totalPrice: Int = 0
+        for (l in listOrdered) {
+            totalPrice += Integer.parseInt(l.Price)
+        }
+
         recyclerView = findViewById(R.id.place_order_recycler_view)
         layoutManager = LinearLayoutManager(this)
         recyclerView.layoutManager = layoutManager
@@ -77,21 +100,48 @@ class placeOrderActivity : AppCompatActivity() {
 
 
         makeOrder.setOnClickListener {
-            val request = Request(
-                Name = Name.text.toString(),
-                Phone = Phone.text.toString(),
-                Address = Address.text.toString(),
-                Dishes = listOrdered
-            )
 
-
-            var mDatabase = FirebaseDatabase.getInstance().reference
-            mDatabase.child("Requests").child(System.currentTimeMillis().toString()).setValue(request)
-object : DBHelper(baseContext){}.deleteAll()
-            startActivity(Intent(this@placeOrderActivity, RequestsListActivity::class.java))
+            var payPalPayment = PayPalPayment(BigDecimal(totalPrice), "USD", "order", PayPalPayment.PAYMENT_INTENT_SALE)
+            var intent = Intent(applicationContext, PaymentActivity::class.java)
+            intent.putExtra(PayPalService.EXTRA_PAYPAL_CONFIGURATION, paypalConfig)
+            intent.putExtra(PaymentActivity.EXTRA_PAYMENT, payPalPayment)
+            startActivityForResult(intent, PAYPAL_REQUEST_CODE)
         }
+    }
 
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        if (requestCode== PAYPAL_REQUEST_CODE){
+            if (resultCode==RESULT_OK){
+                var paymentConfirmation:PaymentConfirmation= data!!.getParcelableExtra(PaymentActivity.EXTRA_RESULT_CONFIRMATION)
+                if (paymentConfirmation != null){
+                    try {
+                        var paymentDetail:String=paymentConfirmation.toJSONObject().toString(4)
 
+                        val request = Request(
+                            Name = Name.text.toString(),
+                            Phone = Phone.text.toString(),
+                            Address = Address.text.toString(),
+                            Status = "Accepted",
+                            Dishes = listOrdered
+                        )
+                        var mDatabase = FirebaseDatabase.getInstance().reference
+                        mDatabase.child("Requests").child(System.currentTimeMillis().toString()).setValue(request)
+                        object : DBHelper(baseContext){}.deleteAll()
+                        Toast.makeText(this,"Thanks ! Your order is made ! ", Toast.LENGTH_LONG).show()
+                    } catch (e: JSONException){
+                        e.printStackTrace()
+                        error("Something wrong")
+                    }
+
+                }
+            }
+
+        }
+        else if (requestCode==Activity.RESULT_CANCELED)
+            Toast.makeText(this,"Payment cancel ", Toast.LENGTH_LONG).show()
+        else if (requestCode==PaymentActivity.RESULT_EXTRAS_INVALID)
+            Toast.makeText(this,"Invalid Payment", Toast.LENGTH_LONG).show()
+       // super.onActivityResult(requestCode, resultCode, data)
     }
